# public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Online Money Transfer Rates API",
    "version": "1.0.0",
    "summary": "Timestamped money transfer rate evidence",
    "description": "Free JSON and CSV access to current and historical corridor comparisons. Published figures must carry a clearly visible link to the matching Online Money Transfer corridor and keep each rate's capture time and evidence status visible.",
    "termsOfService": "https://onlinemoneytransfer.co.uk/api/terms",
    "contact": {
      "name": "Online Money Transfer",
      "url": "https://onlinemoneytransfer.co.uk/api"
    },
    "license": {
      "name": "Free with corridor-specific attribution",
      "url": "https://onlinemoneytransfer.co.uk/api/terms"
    }
  },
  "jsonSchemaDialect": "https://json-schema.org/draft/2020-12/schema",
  "servers": [
    {
      "url": "https://onlinemoneytransfer.co.uk",
      "description": "Production"
    }
  ],
  "security": [],
  "tags": [
    {
      "name": "Rates",
      "description": "Current and historical money transfer rate evidence"
    }
  ],
  "paths": {
    "/api/v1/corridors": {
      "get": {
        "operationId": "listCorridors",
        "summary": "List every supported corridor",
        "description": "Returns the valid route slugs, currency pairs, standard comparison amounts and canonical URLs.",
        "tags": ["Rates"],
        "responses": {
          "200": {
            "description": "Supported corridor catalogue",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CorridorList" }
              }
            }
          }
        }
      },
      "options": {
        "operationId": "optionsCorridors",
        "summary": "Inspect CORS support",
        "tags": ["Rates"],
        "responses": {
          "204": { "description": "CORS preflight accepted" }
        }
      }
    },
    "/api/v1/rates/{route}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/Route"
        }
      ],
      "get": {
        "operationId": "getCorridorRates",
        "summary": "Get corridor rates as JSON",
        "description": "Returns the newest usable record per provider and up to 30 earlier comparison sweeps.",
        "tags": ["Rates"],
        "parameters": [
          {
            "$ref": "#/components/parameters/History"
          }
        ],
        "responses": {
          "200": {
            "description": "Current and historical rate evidence",
            "headers": {
              "Access-Control-Allow-Origin": {
                "description": "The public endpoint permits cross-origin reads.",
                "schema": { "type": "string", "const": "*" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RatesResponse" }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UnknownCorridor"
          }
        }
      },
      "options": {
        "operationId": "optionsCorridorRates",
        "summary": "Inspect CORS support",
        "tags": ["Rates"],
        "responses": {
          "204": { "description": "CORS preflight accepted" }
        }
      }
    },
    "/api/v1/rates/{route}/csv": {
      "parameters": [
        {
          "$ref": "#/components/parameters/Route"
        }
      ],
      "get": {
        "operationId": "downloadCorridorRatesCsv",
        "summary": "Download corridor rates as CSV",
        "tags": ["Rates"],
        "parameters": [
          {
            "$ref": "#/components/parameters/History"
          }
        ],
        "responses": {
          "200": {
            "description": "Current and historical rate evidence as CSV",
            "content": {
              "text/csv": {
                "schema": { "type": "string" }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UnknownCorridor"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "Route": {
        "name": "route",
        "in": "path",
        "required": true,
        "description": "An Online Money Transfer corridor slug.",
        "example": "uk-to-united-states",
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        }
      },
      "History": {
        "name": "history",
        "in": "query",
        "required": false,
        "description": "Number of earlier stored comparison sweeps to return.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 30,
          "default": 14
        }
      }
    },
    "responses": {
      "UnknownCorridor": {
        "description": "The corridor slug is not monitored.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "unknown_corridor" },
          "message": { "type": "string" }
        }
      },
      "Corridor": {
        "type": "object",
        "required": ["route", "fromCountry", "fromCurrency", "toCountry", "toCurrency", "standardTestAmount", "url"],
        "properties": {
          "route": { "type": "string" },
          "fromCountry": { "type": "string" },
          "fromCurrency": { "type": "string", "pattern": "^[A-Z]{3}$" },
          "toCountry": { "type": "string" },
          "toCurrency": { "type": "string", "pattern": "^[A-Z]{3}$" },
          "standardTestAmount": { "type": "number" },
          "url": { "type": "string", "format": "uri" }
        }
      },
      "CorridorSummary": {
        "type": "object",
        "required": ["route", "group", "fromCountry", "fromCode", "fromCurrency", "toCountry", "toCode", "toCurrency", "standardTestAmount", "url", "ratesUrl"],
        "properties": {
          "route": { "type": "string" },
          "group": { "type": "string", "enum": ["from-uk", "to-uk", "major"] },
          "fromCountry": { "type": "string" },
          "fromCode": { "type": "string", "pattern": "^[A-Z]{2}$" },
          "fromCurrency": { "type": "string", "pattern": "^[A-Z]{3}$" },
          "toCountry": { "type": "string" },
          "toCode": { "type": "string", "pattern": "^[A-Z]{2}$" },
          "toCurrency": { "type": "string", "pattern": "^[A-Z]{3}$" },
          "standardTestAmount": { "type": "number" },
          "url": { "type": "string", "format": "uri" },
          "ratesUrl": { "type": "string", "format": "uri" }
        }
      },
      "CorridorList": {
        "type": "object",
        "required": ["apiVersion", "generatedAt", "count", "corridors"],
        "properties": {
          "apiVersion": { "type": "string" },
          "generatedAt": { "type": "string", "format": "date-time" },
          "count": { "type": "integer", "minimum": 1 },
          "corridors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CorridorSummary" }
          }
        }
      },
      "Rate": {
        "type": "object",
        "required": ["id", "provider", "providerSlug", "quoteType", "status", "eligibleForPriceRanking", "sourceAmount", "sourceCurrency", "recipientAmount", "recipientCurrency", "exchangeRate", "feeAmount", "feeCurrency", "fundingMethod", "payoutMethod", "promotion", "capturedAt", "receiptUrl"],
        "properties": {
          "id": { "type": "string" },
          "provider": { "type": "string" },
          "providerSlug": { "type": "string" },
          "quoteType": { "type": "string", "enum": ["verified", "indicative"] },
          "status": { "type": "string", "enum": ["verified", "indicative", "stale"] },
          "eligibleForPriceRanking": { "type": "boolean" },
          "sourceAmount": { "type": "number" },
          "sourceCurrency": { "type": "string" },
          "recipientAmount": { "type": "number" },
          "recipientCurrency": { "type": "string" },
          "exchangeRate": { "type": "number" },
          "feeAmount": { "type": "number" },
          "feeCurrency": { "type": "string" },
          "deliveryEstimate": { "type": ["string", "null"] },
          "fundingMethod": { "type": "string" },
          "payoutMethod": { "type": "string" },
          "pricingBasis": { "type": ["string", "null"] },
          "promotion": { "type": "boolean" },
          "capturedAt": { "type": "string", "format": "date-time" },
          "receiptUrl": { "type": "string", "format": "uri" }
        }
      },
      "Snapshot": {
        "type": "object",
        "required": ["id", "kind", "capturedAt", "rates"],
        "properties": {
          "id": { "type": "string" },
          "kind": { "type": "string", "enum": ["current", "crawl-run"] },
          "capturedAt": { "type": ["string", "null"], "format": "date-time" },
          "rates": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Rate" }
          }
        }
      },
      "UseTerms": {
        "type": "object",
        "required": ["price", "attributionRequired", "requiredLink", "wording", "placement", "timestampRequired", "statusRequired", "context"],
        "properties": {
          "price": { "type": "string", "const": "Free" },
          "attributionRequired": { "type": "boolean", "const": true },
          "requiredLink": { "type": "string", "format": "uri" },
          "wording": { "type": "string" },
          "placement": { "type": "string" },
          "timestampRequired": { "type": "boolean", "const": true },
          "statusRequired": { "type": "boolean", "const": true },
          "context": { "type": "string" }
        }
      },
      "RatesResponse": {
        "type": "object",
        "required": ["apiVersion", "generatedAt", "corridor", "current", "history", "useTerms", "evidencePolicy"],
        "properties": {
          "apiVersion": { "type": "string", "example": "1.0" },
          "generatedAt": { "type": "string", "format": "date-time" },
          "corridor": { "$ref": "#/components/schemas/Corridor" },
          "current": { "$ref": "#/components/schemas/Snapshot" },
          "history": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Snapshot" }
          },
          "useTerms": { "$ref": "#/components/schemas/UseTerms" },
          "evidencePolicy": {
            "type": "object",
            "additionalProperties": { "type": ["string", "boolean"] }
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "Human-readable integration guide",
    "url": "https://onlinemoneytransfer.co.uk/api"
  }
}